test: expand extensionapi unit tests
