+++ internal/extensionapi/serverroute_connection_test.go
@@ -9,26 +9,24 @@ import (
 	"bytes"
 	"encoding/json"
 	"fmt"
-	"net/http"
 	"net/http/httptest"
-	"strings"
 	"testing"
 
-	"github.com/go-logr/logr"
-	pluginapi "github.com/jupyter-infra/jupyter-k8s-plugin/api"
-	"github.com/jupyter-infra/jupyter-k8s-plugin/pluginclient"
 	connectionv1alpha1 "github.com/jupyter-infra/jupyter-k8s/api/connection/v1alpha1"
 	workspacev1alpha1 "github.com/jupyter-infra/jupyter-k8s/api/v1alpha1"
 	"github.com/jupyter-infra/jupyter-k8s/internal/jwt"
+	"github.com/stretchr/testify/require"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apiserver/pkg/authentication/user"
-	"k8s.io/apiserver/pkg/endpoints/request"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	ctrlclient "sigs.k8s.io/controller-runtime/pkg/client/fake"
 )
+
+// Shared test fixtures for the serverroute_connection_*_test.go files: mock
+// signer/validator implementations, a failing io.ReadCloser, and helpers for
+// building an ExtensionServer and driving HandleConnectionCreate.
 
 const testUser = "test-user"
 const testWorkspaceMyWorkspace = "myworkspace"
@@ -37,9 +35,13 @@ const testStrategyWebSocket = "ws-strategy"
 // mockSignerFactory for testing
 type mockSignerFactory struct {
 	signer *mockSigner
+	err    error
 }
 
 func (m *mockSignerFactory) CreateSigner(accessStrategy *workspacev1alpha1.WorkspaceAccessStrategy) (jwt.Signer, error) {
+	if m.err != nil {
+		return nil, m.err
+	}
 	return m.signer, nil
 }
 
@@ -86,1160 +88,59 @@ func (e *badReader) Close() error {
 	return nil
 }
 
-// --- generateBearerTokenURL tests ---
-
-func TestGenerateBearerTokenURL(t *testing.T) {
-	workspace := &workspacev1alpha1.Workspace{
+// availableWorkspaceWithStrategy builds a Public, Available workspace referencing
+// an access strategy, plus that access strategy, for driving HandleConnectionCreate
+// end to end.
+func availableWorkspaceWithStrategy(strategySpec workspacev1alpha1.WorkspaceAccessStrategySpec) (*workspacev1alpha1.Workspace, *workspacev1alpha1.WorkspaceAccessStrategy) {
+	strategy := &workspacev1alpha1.WorkspaceAccessStrategy{
+		ObjectMeta: metav1.ObjectMeta{Name: strategyName, Namespace: namespaceDefault},
+		Spec:       strategySpec,
+	}
+	ws := &workspacev1alpha1.Workspace{
 		ObjectMeta: metav1.ObjectMeta{Name: testWorkspace, Namespace: namespaceDefault},
 		Spec: workspacev1alpha1.WorkspaceSpec{
-			AccessStrategy: &workspacev1alpha1.AccessStrategyRef{Name: testStrategy},
+			AccessType:     AccessTypePublic,
+			AccessStrategy: &workspacev1alpha1.AccessStrategyRef{Name: strategyName, Namespace: namespaceDefault},
+		},
+		Status: workspacev1alpha1.WorkspaceStatus{
+			Conditions: []metav1.Condition{
+				{Type: conditionTypeAvailable, Status: metav1.ConditionTrue, Reason: "Ready",
+					LastTransitionTime: metav1.Now()},
+			},
 		},
 	}
+	return ws, strategy
+}
 
-	accessStrategy := &workspacev1alpha1.WorkspaceAccessStrategy{
-		ObjectMeta: metav1.ObjectMeta{Name: testStrategy, Namespace: namespaceDefault},
-		Spec: workspacev1alpha1.WorkspaceAccessStrategySpec{
-			BearerAuthURLTemplate: "https://test.com/workspaces/{{.Workspace.Namespace}}/{{.Workspace.Name}}/bearer-auth",
-		},
-	}
-
+func newConnectionServer(t *testing.T, objs ...client.Object) *ExtensionServer {
+	t.Helper()
 	scheme := runtime.NewScheme()
-	_ = workspacev1alpha1.AddToScheme(scheme)
-	fakeClient := ctrlclient.NewClientBuilder().WithScheme(scheme).WithObjects(workspace, accessStrategy).Build()
-
-	server := &ExtensionServer{
-		config:        &ExtensionConfig{},
-		signerFactory: &mockSignerFactory{signer: &mockSigner{token: testToken}},
-		k8sClient:     fakeClient,
-	}
-
-	req := httptest.NewRequest("POST", "/test", nil)
-	req.Header.Set("X-Remote-User", testUser)
-
-	url, err := server.generateBearerTokenURL(req, workspace, accessStrategy)
-
-	if err != nil {
-		t.Errorf("unexpected error: %v", err)
-	}
-
-	expected := "https://test.com/workspaces/default/test-workspace/bearer-auth?token=test-token"
-	if url != expected {
-		t.Errorf("expected %s, got %s", expected, url)
-	}
-}
-
-func TestGenerateBearerTokenURL_SubdomainRouting(t *testing.T) {
-	workspace := &workspacev1alpha1.Workspace{
-		ObjectMeta: metav1.ObjectMeta{Name: testWorkspaceMyWorkspace, Namespace: namespaceDefault},
-		Spec: workspacev1alpha1.WorkspaceSpec{
-			AccessStrategy: &workspacev1alpha1.AccessStrategyRef{Name: "subdomain-strategy"},
-		},
-	}
-
-	accessStrategy := &workspacev1alpha1.WorkspaceAccessStrategy{
-		ObjectMeta: metav1.ObjectMeta{Name: "subdomain-strategy", Namespace: namespaceDefault},
-		Spec: workspacev1alpha1.WorkspaceAccessStrategySpec{
-			BearerAuthURLTemplate: "https://{{.Workspace.Name}}-{{b32encode .Workspace.Namespace}}.example.com/bearer-auth",
-		},
-	}
-
-	scheme := runtime.NewScheme()
-	_ = workspacev1alpha1.AddToScheme(scheme)
-	fakeClient := ctrlclient.NewClientBuilder().WithScheme(scheme).WithObjects(workspace, accessStrategy).Build()
-
-	server := &ExtensionServer{
-		config:        &ExtensionConfig{},
-		signerFactory: &mockSignerFactory{signer: &mockSigner{token: testToken}},
-		k8sClient:     fakeClient,
-	}
-
-	req := httptest.NewRequest("POST", "/test", nil)
-	req.Header.Set("X-Remote-User", testUser)
-
-	url, err := server.generateBearerTokenURL(req, workspace, accessStrategy)
-
-	if err != nil {
-		t.Errorf("unexpected error: %v", err)
-	}
-	expected := "https://myworkspace-mrswmylvnr2a.example.com/bearer-auth?token=test-token"
-	if url != expected {
-		t.Errorf("expected %s, got %s", expected, url)
-	}
-}
-
-func TestGenerateBearerTokenURL_PassesGroupsAndExtra(t *testing.T) {
-	workspace := &workspacev1alpha1.Workspace{
-		ObjectMeta: metav1.ObjectMeta{Name: testWorkspace, Namespace: namespaceDefault},
-		Spec: workspacev1alpha1.WorkspaceSpec{
-			AccessStrategy: &workspacev1alpha1.AccessStrategyRef{Name: testStrategy},
-		},
-	}
-
-	accessStrategy := &workspacev1alpha1.WorkspaceAccessStrategy{
-		ObjectMeta: metav1.ObjectMeta{Name: testStrategy, Namespace: namespaceDefault},
-		Spec: workspacev1alpha1.WorkspaceAccessStrategySpec{
-			BearerAuthURLTemplate: "https://test.com/bearer-auth",
-		},
-	}
-
-	scheme := runtime.NewScheme()
-	_ = workspacev1alpha1.AddToScheme(scheme)
-	fakeClient := ctrlclient.NewClientBuilder().WithScheme(scheme).WithObjects(workspace, accessStrategy).Build()
-
-	signer := &mockSigner{token: testToken}
-	server := &ExtensionServer{
-		config:        &ExtensionConfig{},
-		signerFactory: &mockSignerFactory{signer: signer},
-		k8sClient:     fakeClient,
-	}
-
-	req := httptest.NewRequest("POST", "/test", nil)
-	userInfo := &user.DefaultInfo{
-		Name:   testUser,
-		Groups: []string{"cluster-workspace-admin", "system:authenticated"},
-		Extra: map[string][]string{
-			"arn":          {"arn:aws:sts::123456:assumed-role/Admin/session"},
-			"canonicalarn": {"arn:aws:iam::123456:role/Admin"},
-		},
-	}
-	ctx := request.WithUser(req.Context(), userInfo)
-	req = req.WithContext(ctx)
-
-	_, err := server.generateBearerTokenURL(req, workspace, accessStrategy)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	// Verify groups were passed to the signer
-	if len(signer.lastGroups) != 2 {
-		t.Fatalf("expected 2 groups, got %d: %v", len(signer.lastGroups), signer.lastGroups)
-	}
-	if signer.lastGroups[0] != "cluster-workspace-admin" || signer.lastGroups[1] != "system:authenticated" {
-		t.Errorf("unexpected groups: %v", signer.lastGroups)
-	}
-
-	// Verify extra was passed to the signer
-	if len(signer.lastExtra) != 2 {
-		t.Fatalf("expected 2 extra keys, got %d: %v", len(signer.lastExtra), signer.lastExtra)
-	}
-	if signer.lastExtra["arn"][0] != "arn:aws:sts::123456:assumed-role/Admin/session" {
-		t.Errorf("unexpected extra arn: %v", signer.lastExtra["arn"])
-	}
-	if signer.lastExtra["canonicalarn"][0] != "arn:aws:iam::123456:role/Admin" {
-		t.Errorf("unexpected extra canonicalarn: %v", signer.lastExtra["canonicalarn"])
-	}
-}
-
-func TestGenerateBearerTokenURL_NoContextFallsBackToEmptyGroupsAndExtra(t *testing.T) {
-	workspace := &workspacev1alpha1.Workspace{
-		ObjectMeta: metav1.ObjectMeta{Name: testWorkspace, Namespace: namespaceDefault},
-		Spec: workspacev1alpha1.WorkspaceSpec{
-			AccessStrategy: &workspacev1alpha1.AccessStrategyRef{Name: testStrategy},
-		},
-	}
-
-	accessStrategy := &workspacev1alpha1.WorkspaceAccessStrategy{
-		ObjectMeta: metav1.ObjectMeta{Name: testStrategy, Namespace: namespaceDefault},
-		Spec: workspacev1alpha1.WorkspaceAccessStrategySpec{
-			BearerAuthURLTemplate: "https://test.com/bearer-auth",
-		},
-	}
-
-	scheme := runtime.NewScheme()
-	_ = workspacev1alpha1.AddToScheme(scheme)
-	fakeClient := ctrlclient.NewClientBuilder().WithScheme(scheme).WithObjects(workspace, accessStrategy).Build()
-
-	signer := &mockSigner{token: testToken}
-	server := &ExtensionServer{
-		config:        &ExtensionConfig{},
-		signerFactory: &mockSignerFactory{signer: signer},
-		k8sClient:     fakeClient,
-	}
-
-	// Request with header-based user, no Kubernetes context
-	req := httptest.NewRequest("POST", "/test", nil)
-	req.Header.Set("X-Remote-User", testUser)
-
-	_, err := server.generateBearerTokenURL(req, workspace, accessStrategy)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	// Groups and extra should be nil when no k8s context
-	if signer.lastGroups != nil {
-		t.Errorf("expected nil groups, got %v", signer.lastGroups)
-	}
-	if signer.lastExtra != nil {
-		t.Errorf("expected nil extra, got %v", signer.lastExtra)
-	}
-}
-
-func TestGenerateBearerTokenURL_NoAccessStrategy(t *testing.T) {
-	server := &ExtensionServer{
-		config:        &ExtensionConfig{},
-		signerFactory: &mockSignerFactory{signer: &mockSigner{token: testToken}},
-	}
-
-	req := httptest.NewRequest("POST", "/test", nil)
-	req.Header.Set("X-Remote-User", testUser)
-
-	_, err := server.generateBearerTokenURL(req, &workspacev1alpha1.Workspace{}, nil)
-
-	if err == nil {
-		t.Error("expected error for missing AccessStrategy, got nil")
-	}
-	if !strings.Contains(err.Error(), "no AccessStrategy configured") {
-		t.Errorf("expected AccessStrategy error, got: %v", err)
-	}
-}
-
-func TestGenerateBearerTokenURL_MissingTemplate(t *testing.T) {
-	accessStrategy := &workspacev1alpha1.WorkspaceAccessStrategy{
-		Spec: workspacev1alpha1.WorkspaceAccessStrategySpec{
-			BearerAuthURLTemplate: "",
-		},
-	}
-
-	server := &ExtensionServer{
-		config:        &ExtensionConfig{},
-		signerFactory: &mockSignerFactory{signer: &mockSigner{token: testToken}},
-	}
-
-	req := httptest.NewRequest("POST", "/test", nil)
-	req.Header.Set("X-Remote-User", testUser)
-
-	_, err := server.generateBearerTokenURL(req, &workspacev1alpha1.Workspace{}, accessStrategy)
-
-	if err == nil {
-		t.Error("expected error for missing BearerAuthURLTemplate, got nil")
-	}
-	if !strings.Contains(err.Error(), "BearerAuthURLTemplate not configured") {
-		t.Errorf("expected template error, got: %v", err)
-	}
-}
-
-// --- generatePluginConnectionURL tests ---
-
-func TestGeneratePluginConnectionURL_Success(t *testing.T) {
-	workspace := &workspacev1alpha1.Workspace{
-		ObjectMeta: metav1.ObjectMeta{Name: testWorkspace, Namespace: namespaceDefault},
-	}
-
-	// Create httptest server simulating the plugin
-	pluginSrv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		var req pluginapi.CreateSessionRequest
-		if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		}
-		if req.PodUID != "test-uid-123" {
-			t.Errorf("expected podUID test-uid-123, got %s", req.PodUID)
-		}
-		if req.ConnectionContext["ssmDocumentName"] != "test-document" {
-			t.Errorf("expected ssmDocumentName test-document, got %s", req.ConnectionContext["ssmDocumentName"])
-		}
-		if req.ConnectionType != connectionTypeVSCodeRemote {
-			t.Errorf("expected connectionType vscode-remote, got %s", req.ConnectionType)
-		}
-		w.Header().Set("Content-Type", "application/json")
-		_ = json.NewEncoder(w).Encode(pluginapi.CreateSessionResponse{
-			ConnectionURL: "vscode://vscode-remote/ssh-remote+test-workspace/home/user",
-		})
-	}))
-	defer pluginSrv.Close()
-
-	server := &ExtensionServer{
-		config: &ExtensionConfig{},
-		pluginClients: map[string]*pluginclient.PluginClient{
-			pluginNameAWS: pluginclient.NewPluginClient(pluginSrv.URL, logr.Discard()),
-		},
-	}
-
-	resolvedContext := map[string]string{
-		"podUid":          "test-uid-123",
-		"ssmDocumentName": "test-document",
-	}
-
-	req := httptest.NewRequest("POST", "/test", nil)
-	url, err := server.generatePluginConnectionURL(req, workspace, pluginNameAWS, "createSession", connectionTypeVSCodeRemote, resolvedContext, namespaceDefault)
-
-	if err != nil {
-		t.Errorf("unexpected error: %v", err)
-	}
-	if url != "vscode://vscode-remote/ssh-remote+test-workspace/home/user" {
-		t.Errorf("unexpected URL: %s", url)
-	}
-}
-
-func TestGeneratePluginConnectionURL_PluginError(t *testing.T) {
-	workspace := &workspacev1alpha1.Workspace{
-		ObjectMeta: metav1.ObjectMeta{Name: testWorkspace, Namespace: namespaceDefault},
-	}
-
-	pluginSrv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(http.StatusInternalServerError)
-		_ = json.NewEncoder(w).Encode(map[string]string{"error": "SSM creation failed"})
-	}))
-	defer pluginSrv.Close()
-
-	server := &ExtensionServer{
-		config: &ExtensionConfig{},
-		pluginClients: map[string]*pluginclient.PluginClient{
-			pluginNameAWS: pluginclient.NewPluginClient(pluginSrv.URL, logr.Discard()),
-		},
-	}
-
-	req := httptest.NewRequest("POST", "/test", nil)
-	_, err := server.generatePluginConnectionURL(req, workspace, pluginNameAWS, "createSession", connectionTypeVSCodeRemote, map[string]string{}, namespaceDefault)
-
-	if err == nil {
-		t.Error("expected error from plugin client")
-	}
-}
-
-func TestGeneratePluginConnectionURL_NoPlugin(t *testing.T) {
-	server := &ExtensionServer{
-		config:        &ExtensionConfig{},
-		pluginClients: map[string]*pluginclient.PluginClient{},
-	}
-
-	req := httptest.NewRequest("POST", "/test", nil)
-	_, err := server.generatePluginConnectionURL(req, &workspacev1alpha1.Workspace{}, pluginNameAWS, "createSession", connectionTypeVSCodeRemote, map[string]string{}, namespaceDefault)
-
-	if err == nil {
-		t.Error("expected error for missing plugin")
-	}
-	if !strings.Contains(err.Error(), "no plugin endpoint configured") {
-		t.Errorf("expected plugin error, got: %v", err)
-	}
-}
-
-func TestGeneratePluginConnectionURL_UnsupportedAction(t *testing.T) {
-	server := &ExtensionServer{
-		config:        &ExtensionConfig{},
-		pluginClients: map[string]*pluginclient.PluginClient{pluginNameAWS: pluginclient.NewPluginClient("http://localhost:8080", logr.Discard())},
-	}
-
-	req := httptest.NewRequest("POST", "/test", nil)
-	_, err := server.generatePluginConnectionURL(req, &workspacev1alpha1.Workspace{}, pluginNameAWS, "unknownAction", connectionTypeVSCodeRemote, map[string]string{}, namespaceDefault)
-
-	if err == nil {
-		t.Error("expected error for unsupported action")
-	}
-	if !strings.Contains(err.Error(), "unsupported plugin action") {
-		t.Errorf("expected unsupported action error, got: %v", err)
-	}
-}
-
-// --- resolveConnectionHandler tests ---
-
-func TestResolveConnectionHandler(t *testing.T) {
-	tests := []struct {
-		name           string
-		accessStrategy *workspacev1alpha1.WorkspaceAccessStrategy
-		connectionType string
-		expectedPlugin string
-		expectedAction string
-		expectedFound  bool
-	}{
-		{
-			name:           "nil access strategy",
-			accessStrategy: nil,
-			connectionType: connectionTypeWebUI,
-			expectedFound:  false,
-		},
-		{
-			name: "found in handler map",
-			accessStrategy: &workspacev1alpha1.WorkspaceAccessStrategy{
-				Spec: workspacev1alpha1.WorkspaceAccessStrategySpec{
-					CreateConnectionHandlerMap: map[string]string{
-						connectionTypeVSCodeRemote: "aws:createSession",
-					},
-					CreateConnectionHandler: handlerK8sNative,
-				},
-			},
-			connectionType: connectionTypeVSCodeRemote,
-			expectedPlugin: pluginNameAWS,
-			expectedAction: "createSession",
-			expectedFound:  true,
-		},
-		{
-			name: "falls back to default handler",
-			accessStrategy: &workspacev1alpha1.WorkspaceAccessStrategy{
-				Spec: workspacev1alpha1.WorkspaceAccessStrategySpec{
-					CreateConnectionHandlerMap: map[string]string{
-						connectionTypeVSCodeRemote: "aws:createSession",
-					},
-					CreateConnectionHandler: handlerK8sNative,
-				},
-			},
-			connectionType: connectionTypeWebUI,
-			expectedPlugin: handlerK8sNative,
-			expectedAction: "",
-			expectedFound:  true,
-		},
-		{
-			name: "no handler configured",
-			accessStrategy: &workspacev1alpha1.WorkspaceAccessStrategy{
-				Spec: workspacev1alpha1.WorkspaceAccessStrategySpec{},
-			},
-			connectionType: connectionTypeWebUI,
-			expectedFound:  false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			plugin, action, found := resolveConnectionHandler(tt.accessStrategy, tt.connectionType)
-			if found != tt.expectedFound {
-				t.Errorf("expected found=%v, got %v", tt.expectedFound, found)
-			}
-			if plugin != tt.expectedPlugin {
-				t.Errorf("expected plugin=%q, got %q", tt.expectedPlugin, plugin)
-			}
-			if action != tt.expectedAction {
-				t.Errorf("expected action=%q, got %q", tt.expectedAction, action)
-			}
-		})
-	}
-}
-
-// --- validateConnection tests ---
-
-func TestValidateConnection(t *testing.T) {
-	scheme := runtime.NewScheme()
-	_ = workspacev1alpha1.AddToScheme(scheme)
-	_ = corev1.AddToScheme(scheme)
-
-	tests := []struct {
-		name               string
-		workspace          *workspacev1alpha1.Workspace
-		accessStrategy     *workspacev1alpha1.WorkspaceAccessStrategy
-		expectedStatusCode int
-		expectedError      string
-	}{
-		{
-			name: "workspace not available",
-			workspace: &workspacev1alpha1.Workspace{
-				ObjectMeta: metav1.ObjectMeta{Name: "ws", Namespace: namespaceDefault},
-				Spec: workspacev1alpha1.WorkspaceSpec{
-					AccessStrategy: &workspacev1alpha1.AccessStrategyRef{Name: "as"},
-				},
-				Status: workspacev1alpha1.WorkspaceStatus{
-					Conditions: []metav1.Condition{{Type: conditionTypeAvailable, Status: metav1.ConditionFalse}},
-				},
-			},
-			accessStrategy: &workspacev1alpha1.WorkspaceAccessStrategy{
-				ObjectMeta: metav1.ObjectMeta{Name: "as", Namespace: namespaceDefault},
-				Spec: workspacev1alpha1.WorkspaceAccessStrategySpec{
-					CreateConnectionHandler: handlerK8sNative,
-					BearerAuthURLTemplate:   "https://example.com",
-				},
-			},
-			expectedStatusCode: http.StatusBadRequest,
-			expectedError:      "workspace is not available",
-		},
-		{
-			name: "validation passes",
-			workspace: &workspacev1alpha1.Workspace{
-				ObjectMeta: metav1.ObjectMeta{Name: "ws", Namespace: namespaceDefault},
-				Spec: workspacev1alpha1.WorkspaceSpec{
-					AccessStrategy: &workspacev1alpha1.AccessStrategyRef{Name: "as"},
-				},
-				Status: workspacev1alpha1.WorkspaceStatus{
-					Conditions: []metav1.Condition{{Type: conditionTypeAvailable, Status: metav1.ConditionTrue}},
-				},
-			},
-			accessStrategy: &workspacev1alpha1.WorkspaceAccessStrategy{
-				ObjectMeta: metav1.ObjectMeta{Name: "as", Namespace: namespaceDefault},
-				Spec: workspacev1alpha1.WorkspaceAccessStrategySpec{
-					CreateConnectionHandler: handlerK8sNative,
-					BearerAuthURLTemplate:   "https://example.com",
-					CreateConnectionContext: map[string]string{
-						"staticKey": "staticValue",
-					},
-				},
-			},
-			expectedStatusCode: 0,
-			expectedError:      "",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			var objects []client.Object
-			if tt.workspace != nil {
-				objects = append(objects, tt.workspace)
-			}
-			if tt.accessStrategy != nil {
-				objects = append(objects, tt.accessStrategy)
-			}
-
-			fakeClient := ctrlclient.NewClientBuilder().WithScheme(scheme).WithObjects(objects...).Build()
-
-			server := &ExtensionServer{
-				k8sClient: fakeClient,
-			}
-
-			logger := ctrl.Log.WithName("test")
-			_, resolvedCtx, statusCode, err := server.validateConnection(tt.workspace, logger)
-
-			if statusCode != tt.expectedStatusCode {
-				t.Errorf("expected status code %d, got %d", tt.expectedStatusCode, statusCode)
-			}
-
-			if tt.expectedError == "" {
-				if err != nil {
-					t.Errorf("expected no error, got %v", err)
-				}
-				// Check that context was returned when validation passes
-				if resolvedCtx == nil && tt.accessStrategy != nil && len(tt.accessStrategy.Spec.CreateConnectionContext) > 0 {
-					t.Error("expected resolved context to be returned")
-				}
-			} else {
-				if err == nil {
-					t.Errorf("expected error containing %q, got nil", tt.expectedError)
-				} else if !strings.Contains(err.Error(), tt.expectedError) {
-					t.Errorf("expected error containing %q, got %q", tt.expectedError, err.Error())
-				}
-			}
-		})
-	}
-}
-
-// --- HandleConnectionCreate tests ---
-
-func TestHandleConnectionCreateValidation(t *testing.T) {
-	server := &ExtensionServer{
-		config: &ExtensionConfig{},
-	}
-
-	tests := []struct {
-		name           string
-		method         string
-		path           string
-		body           interface{}
-		expectedStatus int
-	}{
-		{
-			name:           "wrong method",
-			method:         "GET",
-			path:           connectionsPath,
-			body:           nil,
-			expectedStatus: http.StatusBadRequest,
-		},
-		{
-			name:           "invalid path",
-			method:         http.MethodPost,
-			path:           "/invalid/path",
-			body:           nil,
-			expectedStatus: http.StatusBadRequest,
-		},
-		{
-			name:           "invalid JSON",
-			method:         http.MethodPost,
-			path:           connectionsPath,
-			body:           "invalid json",
-			expectedStatus: http.StatusBadRequest,
-		},
-		{
-			name:   "missing workspace name",
-			method: http.MethodPost,
-			path:   connectionsPath,
-			body: connectionv1alpha1.WorkspaceConnectionRequest{
-				Spec: connectionv1alpha1.WorkspaceConnectionRequestSpec{
-					WorkspaceConnectionType: connectionTypeVSCodeRemote,
-				},
-			},
-			expectedStatus: http.StatusBadRequest,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			var bodyBytes []byte
-			if tt.body != nil {
-				if str, ok := tt.body.(string); ok {
-					bodyBytes = []byte(str)
-				} else {
-					bodyBytes, _ = json.Marshal(tt.body)
-				}
-			}
-
-			req := httptest.NewRequest(tt.method, tt.path, bytes.NewReader(bodyBytes))
-			w := httptest.NewRecorder()
-
-			server.HandleConnectionCreate(w, req)
-
-			if w.Code != tt.expectedStatus {
-				t.Errorf("expected status %d, got %d", tt.expectedStatus, w.Code)
-			}
-		})
-	}
-}
-
-func TestHandleConnectionCreateReadBodyError(t *testing.T) {
-	server := &ExtensionServer{
-		config: &ExtensionConfig{},
-	}
-
-	req := httptest.NewRequest("POST", connectionsPath, nil)
-	req.Body = &badReader{}
-	w := httptest.NewRecorder()
-
-	server.HandleConnectionCreate(w, req)
-
-	if w.Code != http.StatusBadRequest {
-		t.Errorf("expected status %d for body read error, got %d", http.StatusBadRequest, w.Code)
-	}
-}
-
-func TestHandleConnectionCreateInvalidConnectionType(t *testing.T) {
-	server := &ExtensionServer{
-		config: &ExtensionConfig{},
-	}
-
-	reqBody := connectionv1alpha1.WorkspaceConnectionRequest{
-		Spec: connectionv1alpha1.WorkspaceConnectionRequestSpec{
-			WorkspaceName:           testWorkspace,
-			WorkspaceConnectionType: invalidConnectionType,
-		},
-	}
-
-	bodyBytes, _ := json.Marshal(reqBody)
-	httpReq := httptest.NewRequest("POST", connectionsPath, bytes.NewReader(bodyBytes))
-	w := httptest.NewRecorder()
-
-	server.HandleConnectionCreate(w, httpReq)
-
-	if w.Code != http.StatusBadRequest {
-		t.Errorf("expected status %d for invalid connection type, got %d", http.StatusBadRequest, w.Code)
-	}
-}
-
-func TestHandleConnectionCreateAuthorizationError(t *testing.T) {
-	server := &ExtensionServer{
-		config: &ExtensionConfig{},
-	}
-
-	reqBody := connectionv1alpha1.WorkspaceConnectionRequest{
-		Spec: connectionv1alpha1.WorkspaceConnectionRequestSpec{
-			WorkspaceName:           testWorkspace,
-			WorkspaceConnectionType: connectionv1alpha1.ConnectionTypeWebUI,
-		},
-	}
-
-	bodyBytes, _ := json.Marshal(reqBody)
-	httpReq := httptest.NewRequest("POST", connectionsPath, bytes.NewReader(bodyBytes))
-	w := httptest.NewRecorder()
-
-	server.HandleConnectionCreate(w, httpReq)
-
-	if w.Code != http.StatusInternalServerError {
-		t.Errorf("expected status %d for authorization error, got %d", http.StatusInternalServerError, w.Code)
-	}
-}
-
-func TestHandleConnectionCreateInvalidMethod(t *testing.T) {
-	server := &ExtensionServer{
-		config: &ExtensionConfig{},
-	}
-
-	httpReq := httptest.NewRequest("GET", connectionsPath, nil)
-	w := httptest.NewRecorder()
-
-	server.HandleConnectionCreate(w, httpReq)
-
-	if w.Code != http.StatusBadRequest {
-		t.Errorf("expected status %d for invalid method, got %d", http.StatusBadRequest, w.Code)
-	}
-}
-
-func TestHandleConnectionCreateWebUIPath(t *testing.T) {
-	scheme := runtime.NewScheme()
-	_ = corev1.AddToScheme(scheme)
+	require.NoError(t, corev1.AddToScheme(scheme))
+	require.NoError(t, workspacev1alpha1.AddToScheme(scheme))
 	logger := ctrl.Log.WithName("test")
-
-	server := &ExtensionServer{
-		config:    &ExtensionConfig{},
-		k8sClient: ctrlclient.NewClientBuilder().WithScheme(scheme).Build(),
-		logger:    &logger,
-	}
-
-	reqBody := connectionv1alpha1.WorkspaceConnectionRequest{
-		Spec: connectionv1alpha1.WorkspaceConnectionRequestSpec{
-			WorkspaceName:           "test",
-			WorkspaceConnectionType: connectionv1alpha1.ConnectionTypeWebUI,
-		},
-	}
-
-	bodyBytes, _ := json.Marshal(reqBody)
-	httpReq := httptest.NewRequest("POST", connectionsPath, bytes.NewReader(bodyBytes))
-	httpReq.Header.Set("X-User", "test")
-	w := httptest.NewRecorder()
-
-	server.HandleConnectionCreate(w, httpReq)
-	// Covers WebUI path regardless of final status
-}
-
-func TestHandleConnectionCreateWithWorkspace(t *testing.T) {
-	scheme := runtime.NewScheme()
-	_ = corev1.AddToScheme(scheme)
-	_ = workspacev1alpha1.AddToScheme(scheme)
-
-	workspace := &workspacev1alpha1.Workspace{
-		ObjectMeta: metav1.ObjectMeta{Name: testWorkspace, Namespace: namespaceDefault},
-		Spec: workspacev1alpha1.WorkspaceSpec{
-			AccessType: "Public",
-		},
-	}
-
-	fakeClient := ctrlclient.NewClientBuilder().WithScheme(scheme).WithObjects(workspace).Build()
-	logger := ctrl.Log.WithName("test")
-
-	server := &ExtensionServer{
+	return &ExtensionServer{
 		config:        &ExtensionConfig{},
-		k8sClient:     fakeClient,
+		k8sClient:     ctrlclient.NewClientBuilder().WithScheme(scheme).WithObjects(objs...).Build(),
 		logger:        &logger,
 		signerFactory: &mockSignerFactory{signer: &mockSigner{token: testToken}},
 	}
+}
 
+func postConnection(t *testing.T, server *ExtensionServer, connectionType string) *httptest.ResponseRecorder {
+	t.Helper()
 	reqBody := connectionv1alpha1.WorkspaceConnectionRequest{
 		Spec: connectionv1alpha1.WorkspaceConnectionRequestSpec{
 			WorkspaceName:           testWorkspace,
-			WorkspaceConnectionType: connectionv1alpha1.ConnectionTypeWebUI,
+			WorkspaceConnectionType: connectionType,
 		},
 	}
+	bodyBytes, err := json.Marshal(reqBody)
+	require.NoError(t, err)
 
-	bodyBytes, _ := json.Marshal(reqBody)
 	httpReq := httptest.NewRequest("POST", connectionsPath, bytes.NewReader(bodyBytes))
-	httpReq.Header.Set("X-User", "test-user")
+	httpReq.Header.Set("X-User", testUser)
 	w := httptest.NewRecorder()
 
 	server.HandleConnectionCreate(w, httpReq)
-	// Should pass authorization and reach validation
-}
-
-// --- Other helper tests ---
-
-func TestValidateWorkspaceConnectionRequest(t *testing.T) {
-	tests := []struct {
-		name        string
-		req         *connectionv1alpha1.WorkspaceConnectionRequest
-		expectError bool
-		errorMsg    string
-	}{
-		{
-			name: "valid vscode request",
-			req: &connectionv1alpha1.WorkspaceConnectionRequest{
-				Spec: connectionv1alpha1.WorkspaceConnectionRequestSpec{
-					WorkspaceName:           testWorkspace,
-					WorkspaceConnectionType: connectionTypeVSCodeRemote,
-				},
-			},
-		},
-		{
-			name: "valid web-ui request",
-			req: &connectionv1alpha1.WorkspaceConnectionRequest{
-				Spec: connectionv1alpha1.WorkspaceConnectionRequestSpec{
-					WorkspaceName:           testWorkspace,
-					WorkspaceConnectionType: connectionv1alpha1.ConnectionTypeWebUI,
-				},
-			},
-		},
-		{
-			name: "valid kiro-remote request",
-			req: &connectionv1alpha1.WorkspaceConnectionRequest{
-				Spec: connectionv1alpha1.WorkspaceConnectionRequestSpec{
-					WorkspaceName:           testWorkspace,
-					WorkspaceConnectionType: "kiro-remote",
-				},
-			},
-		},
-		{
-			name: "valid cursor-remote request",
-			req: &connectionv1alpha1.WorkspaceConnectionRequest{
-				Spec: connectionv1alpha1.WorkspaceConnectionRequestSpec{
-					WorkspaceName:           testWorkspace,
-					WorkspaceConnectionType: "cursor-remote",
-				},
-			},
-		},
-		{
-			name: "valid unknown *-remote request",
-			req: &connectionv1alpha1.WorkspaceConnectionRequest{
-				Spec: connectionv1alpha1.WorkspaceConnectionRequestSpec{
-					WorkspaceName:           testWorkspace,
-					WorkspaceConnectionType: "windsurf-remote",
-				},
-			},
-		},
-		{
-			name: "missing workspace name",
-			req: &connectionv1alpha1.WorkspaceConnectionRequest{
-				Spec: connectionv1alpha1.WorkspaceConnectionRequestSpec{
-					WorkspaceConnectionType: connectionTypeVSCodeRemote,
-				},
-			},
-			expectError: true,
-			errorMsg:    "workspaceName is required",
-		},
-		{
-			name: "missing connection type",
-			req: &connectionv1alpha1.WorkspaceConnectionRequest{
-				Spec: connectionv1alpha1.WorkspaceConnectionRequestSpec{
-					WorkspaceName: testWorkspace,
-				},
-			},
-			expectError: true,
-			errorMsg:    "workspaceConnectionType is required",
-		},
-		{
-			name: "invalid connection type - no -remote suffix",
-			req: &connectionv1alpha1.WorkspaceConnectionRequest{
-				Spec: connectionv1alpha1.WorkspaceConnectionRequestSpec{
-					WorkspaceName:           testWorkspace,
-					WorkspaceConnectionType: invalidConnectionType,
-				},
-			},
-			expectError: true,
-			errorMsg:    "invalid workspaceConnectionType: 'invalid-type'. Must be 'web-ui', 'ssh-over-websocket', or follow the '{ide}-remote' pattern (e.g. 'vscode-remote', 'kiro-remote', 'cursor-remote')",
-		},
-		{
-			name: "invalid connection type - bare remote",
-			req: &connectionv1alpha1.WorkspaceConnectionRequest{
-				Spec: connectionv1alpha1.WorkspaceConnectionRequestSpec{
-					WorkspaceName:           testWorkspace,
-					WorkspaceConnectionType: "-remote",
-				},
-			},
-			expectError: true,
-			errorMsg:    "invalid workspaceConnectionType: '-remote'. Must be 'web-ui', 'ssh-over-websocket', or follow the '{ide}-remote' pattern (e.g. 'vscode-remote', 'kiro-remote', 'cursor-remote')",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := validateWorkspaceConnectionRequest(tt.req)
-
-			if tt.expectError {
-				if err == nil {
-					t.Errorf("expected error but got none")
-				} else if err.Error() != tt.errorMsg {
-					t.Errorf("expected error %q, got %q", tt.errorMsg, err.Error())
-				}
-			} else {
-				if err != nil {
-					t.Errorf("unexpected error: %v", err)
-				}
-			}
-		})
-	}
-}
-
-func TestCheckWorkspaceAuthorizationMissingUser(t *testing.T) {
-	server := &ExtensionServer{
-		config: &ExtensionConfig{},
-	}
-
-	req := httptest.NewRequest("POST", "/test", nil)
-
-	_, result, err := server.checkWorkspaceAuthorization(req, testWorkspace, namespaceDefault)
-
-	if err == nil {
-		t.Error("expected error when user headers are missing")
-	}
-	if result != nil {
-		t.Error("expected nil result when user headers are missing")
-	}
-	expectedMsg := "user not found in request headers"
-	if err.Error() != expectedMsg {
-		t.Errorf("expected error %q, got %q", expectedMsg, err.Error())
-	}
-}
-
-func TestGetUserFromHeaders(t *testing.T) {
-	req := httptest.NewRequest("GET", "/test", nil)
-	req.Header.Set("X-Remote-User", testUser)
-
-	username := GetUser(req)
-	if username != testUser {
-		t.Errorf("expected %s, got %s", testUser, username)
-	}
-}
-
-func TestHasWebUIEnabled(t *testing.T) {
-	tests := []struct {
-		name           string
-		accessStrategy *workspacev1alpha1.WorkspaceAccessStrategy
-		expected       bool
-	}{
-		{
-			name:           "nil access strategy",
-			accessStrategy: nil,
-			expected:       false,
-		},
-		{
-			name: "empty BearerAuthURLTemplate",
-			accessStrategy: &workspacev1alpha1.WorkspaceAccessStrategy{
-				Spec: workspacev1alpha1.WorkspaceAccessStrategySpec{
-					BearerAuthURLTemplate: "",
-				},
-			},
-			expected: false,
-		},
-		{
-			name: "BearerAuthURLTemplate configured",
-			accessStrategy: &workspacev1alpha1.WorkspaceAccessStrategy{
-				Spec: workspacev1alpha1.WorkspaceAccessStrategySpec{
-					BearerAuthURLTemplate: "https://example.com/bearer-auth",
-				},
-			},
-			expected: true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			result := hasWebUIEnabled(tt.accessStrategy)
-			if result != tt.expected {
-				t.Errorf("expected %v, got %v", tt.expected, result)
-			}
-		})
-	}
-}
-
-func TestIsWorkspaceAvailable(t *testing.T) {
-	tests := []struct {
-		name      string
-		workspace *workspacev1alpha1.Workspace
-		expected  bool
-	}{
-		{
-			name: "no conditions",
-			workspace: &workspacev1alpha1.Workspace{
-				Status: workspacev1alpha1.WorkspaceStatus{
-					Conditions: []metav1.Condition{},
-				},
-			},
-			expected: false,
-		},
-		{
-			name: "Available condition is True",
-			workspace: &workspacev1alpha1.Workspace{
-				Status: workspacev1alpha1.WorkspaceStatus{
-					Conditions: []metav1.Condition{
-						{Type: conditionTypeAvailable, Status: metav1.ConditionTrue},
-					},
-				},
-			},
-			expected: true,
-		},
-		{
-			name: "Available condition is False",
-			workspace: &workspacev1alpha1.Workspace{
-				Status: workspacev1alpha1.WorkspaceStatus{
-					Conditions: []metav1.Condition{
-						{Type: conditionTypeAvailable, Status: metav1.ConditionFalse},
-					},
-				},
-			},
-			expected: false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			result := isWorkspaceAvailable(tt.workspace)
-			if result != tt.expected {
-				t.Errorf("expected %v, got %v", tt.expected, result)
-			}
-		})
-	}
-}
-
-func TestIsRemoteConnectionType(t *testing.T) {
-	tests := []struct {
-		connectionType string
-		expected       bool
-	}{
-		{connectionTypeVSCodeRemote, true},
-		{"kiro-remote", true},
-		{"cursor-remote", true},
-		{"windsurf-remote", true},
-		{"my-vscode-remote", true},
-		{connectionTypeWebUI, false},
-		{invalidConnectionType, false},
-		{"-remote", false},
-		{"remote", false},
-		{"", false},
-		{"my--vscode-remote", false},
-		{"vscode-remote-ssh", false},
-		{"-vscode-remote", false},
-		{"vscode remote", false},
-		{"vscode_remote", false},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.connectionType, func(t *testing.T) {
-			result := isRemoteConnectionType(tt.connectionType)
-			if result != tt.expected {
-				t.Errorf("isRemoteConnectionType(%q) = %v, want %v", tt.connectionType, result, tt.expected)
-			}
-		})
-	}
-}
-
-// --- generateWebSocketConnectionURL tests ---
-
-func TestGenerateWebSocketConnectionURL_Success(t *testing.T) {
-	workspace := &workspacev1alpha1.Workspace{
-		ObjectMeta: metav1.ObjectMeta{Name: testWorkspaceMyWorkspace, Namespace: namespaceDefault},
-		Spec: workspacev1alpha1.WorkspaceSpec{
-			AccessStrategy: &workspacev1alpha1.AccessStrategyRef{Name: testStrategyWebSocket},
-		},
-	}
-
-	accessStrategy := &workspacev1alpha1.WorkspaceAccessStrategy{
-		ObjectMeta: metav1.ObjectMeta{Name: testStrategyWebSocket, Namespace: namespaceDefault},
-		Spec: workspacev1alpha1.WorkspaceAccessStrategySpec{
-			BearerAuthURLTemplate: "https://myworkspace-default.example.com/ssh-ws",
-		},
-	}
-
-	server := &ExtensionServer{
-		config:        &ExtensionConfig{},
-		signerFactory: &mockSignerFactory{signer: &mockSigner{token: testToken}},
-	}
-
-	req := httptest.NewRequest("POST", "/test", nil)
-	req.Header.Set("X-Remote-User", testUser)
-
-	connURL, err := server.generateWebSocketConnectionURL(req, workspace, accessStrategy)
-
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	if !strings.HasPrefix(connURL, "wss://") {
-		t.Errorf("expected wss:// scheme, got: %s", connURL)
-	}
-	if !strings.Contains(connURL, "token="+testToken) {
-		t.Errorf("expected token in URL, got: %s", connURL)
-	}
-	if !strings.Contains(connURL, "myworkspace-default.example.com") {
-		t.Errorf("expected host in URL, got: %s", connURL)
-	}
-}
-
-func TestGenerateWebSocketConnectionURL_StripsBearerAuth(t *testing.T) {
-	workspace := &workspacev1alpha1.Workspace{
-		ObjectMeta: metav1.ObjectMeta{Name: testWorkspaceMyWorkspace, Namespace: namespaceDefault},
-		Spec: workspacev1alpha1.WorkspaceSpec{
-			AccessStrategy: &workspacev1alpha1.AccessStrategyRef{Name: testStrategyWebSocket},
-		},
-	}
-
-	// Template with /bearer-auth suffix (shared with web UI)
-	accessStrategy := &workspacev1alpha1.WorkspaceAccessStrategy{
-		ObjectMeta: metav1.ObjectMeta{Name: testStrategyWebSocket, Namespace: namespaceDefault},
-		Spec: workspacev1alpha1.WorkspaceAccessStrategySpec{
-			BearerAuthURLTemplate: "https://myworkspace-default.example.com/bearer-auth",
-		},
-	}
-
-	server := &ExtensionServer{
-		config:        &ExtensionConfig{},
-		signerFactory: &mockSignerFactory{signer: &mockSigner{token: testToken}},
-	}
-
-	req := httptest.NewRequest("POST", "/test", nil)
-	req.Header.Set("X-Remote-User", testUser)
-
-	url, err := server.generateWebSocketConnectionURL(req, workspace, accessStrategy)
-
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	// Should strip /bearer-auth from the URL
-	if strings.Contains(url, "/bearer-auth") {
-		t.Errorf("expected /bearer-auth to be stripped, got: %s", url)
-	}
-	if !strings.HasPrefix(url, "wss://") {
-		t.Errorf("expected wss:// scheme, got: %s", url)
-	}
-}
-
-func TestGenerateWebSocketConnectionURL_NoAccessStrategy(t *testing.T) {
-	workspace := &workspacev1alpha1.Workspace{
-		ObjectMeta: metav1.ObjectMeta{Name: testWorkspaceMyWorkspace, Namespace: namespaceDefault},
-	}
-
-	server := &ExtensionServer{
-		config:        &ExtensionConfig{},
-		signerFactory: &mockSignerFactory{signer: &mockSigner{token: testToken}},
-	}
-
-	req := httptest.NewRequest("POST", "/test", nil)
-	req.Header.Set("X-Remote-User", testUser)
-
-	_, err := server.generateWebSocketConnectionURL(req, workspace, nil)
-
-	if err == nil {
-		t.Error("expected error for nil access strategy")
-	}
-}
-
-func TestGenerateWebSocketConnectionURL_MissingUser(t *testing.T) {
-	workspace := &workspacev1alpha1.Workspace{
-		ObjectMeta: metav1.ObjectMeta{Name: testWorkspaceMyWorkspace, Namespace: namespaceDefault},
-	}
-
-	accessStrategy := &workspacev1alpha1.WorkspaceAccessStrategy{
-		ObjectMeta: metav1.ObjectMeta{Name: testStrategyWebSocket, Namespace: namespaceDefault},
-		Spec: workspacev1alpha1.WorkspaceAccessStrategySpec{
-			BearerAuthURLTemplate: "https://myworkspace-default.example.com/ssh-ws",
-		},
-	}
-
-	server := &ExtensionServer{
-		config:        &ExtensionConfig{},
-		signerFactory: &mockSignerFactory{signer: &mockSigner{token: testToken}},
-	}
-
-	req := httptest.NewRequest("POST", "/test", nil)
-	// No X-Remote-User header
-
-	_, err := server.generateWebSocketConnectionURL(req, workspace, accessStrategy)
-
-	if err == nil {
-		t.Error("expected error for missing user")
-	}
-}
-
-func TestGenerateWebSocketConnectionURL_MissingTemplate(t *testing.T) {
-	workspace := &workspacev1alpha1.Workspace{
-		ObjectMeta: metav1.ObjectMeta{Name: testWorkspaceMyWorkspace, Namespace: namespaceDefault},
-	}
-
-	accessStrategy := &workspacev1alpha1.WorkspaceAccessStrategy{
-		ObjectMeta: metav1.ObjectMeta{Name: testStrategyWebSocket, Namespace: namespaceDefault},
-		Spec:       workspacev1alpha1.WorkspaceAccessStrategySpec{},
-	}
-
-	server := &ExtensionServer{
-		config:        &ExtensionConfig{},
-		signerFactory: &mockSignerFactory{signer: &mockSigner{token: testToken}},
-	}
-
-	req := httptest.NewRequest("POST", "/test", nil)
-	req.Header.Set("X-Remote-User", testUser)
-
-	_, err := server.generateWebSocketConnectionURL(req, workspace, accessStrategy)
-
-	if err == nil {
-		t.Error("expected error for missing BearerAuthURLTemplate")
-	}
+	return w
 }
